added equivalent glDrawElements call
also modified the update fn of the objrenderer to add the appropriate
drawers, in order to issue the necessary glElement drawcalls.

This is actually a pretty nice example for how to not use this library
and the code shows. needing to add drawers in the update-fn is ugly. and
now we have multiple drawers, drawing the same thing, but different
parts. That is really obscure and not in the spirit of rather
declarative alternatives. A better solution would be subpass drawers
or a glMultiDrawElements equivalent. Both are not yet implemented.