switched view and projection matrix
also added another debugging triangle.
Somehow the loaded obj does not show up.  However I _have_ seen a part of
the model somewhere with the correct red diffuse and specular
highlights.  This means the model is either not correctly loaded or the
transformations are still off. However now we can rule out view and
projection (and viewprojection).