fixed the obj-renderer demo
after a LOT of error checking, some thing have changed: I found out,
that the buf-fn didn't work when presented with existing buffer types.
This is now fixed. There were errors in the obj renderer regarding the
use of the load-value-to-array fn. This fn loads vectors into FLOAT
arrays. This is an easy to forget thing, as vectors were meant to
represent glsl vecn types. Since load-value-to-array does actually do
much more than this, maybe a new name should be chosen like:
prepare-for-gpu-upload or something more catchy. Also the main bug of
why the model didn't show up at all was a 0 as the w-component of the
position vector. This is quite embarrasing but it was hidden in plain
sight and should be easy to catch. There has to be an automated way of
spotting something like this. Maybe there should be a bughunt mode as a
feature of the library that will inject on fns and check clojure specs
or an extensible and user-definable "caveats" list.

TODO: clean up