implemented buf-as-texture return type
Actually its a very hacky solution that needs refactoring once the whole
to-primitives refactoring comes into play.

for now: a buffer texture holding :gvecN values is a gsamplerBuffer with
N-1 whitespaces appended. So a float-buffer-texture has a glsl-type of
"samplerBuffer", but a ivec3-buffer-texture has a glsl-type of
"isamplerBuffer  ". This enables to still dispatch and discriminate
based on the glsl-type but make no diference when the glsl-type is used
for shader generation.

I have opted out at this point to use keywords, as usual, since you
normally don't encounter keywords, that contain spaces.