added keyboard input
after some tinkering and micro-benching it seems, that an
concurrentqueue is the best option for semi-low-latency input stuff. I
really don't want each and every keypress to generate an atom-swap. So
that is why they get aggregated in a queue and processed at the next
update-thread loop.

This way the rendering itself is not affected. The rendering thread will
not be touched by this.