mouse input & priority updates
Added mouse input analougous to the keyboard input.
There is one exception: The mouse position.
Having a callback for the mouse movement seemed like recording excessive data.
The idea came up to only poll the mouse position right before the state
gets updated, so the supplied update-fns from the user would see only
the most relevant mouse positions. Those, that coincide with the actual
update-time. The problem here is, that when update-fns are completely
unordered, update-fns might see the mouse-position from the last update,
instead of the current one. This would cause an inherent variation of
the accuracy and lag of the mouse-position.

Therefore a priority update-fn list has been implemented whose fns will
be called before the regular ones. These are also always in a list so
their order is fixed, which should lead to more reproducible results.

The add-prio-update-fn! function might go private in the future. I do
not know yet if this is a good idea.