performance improvements
Used async-clj-profiler to find obvious problems in the code. Turns out
there was a lot of Reflector stuf happening. After some experimenting
with the import-static-all macro I came to the conclusion, that for now
the best option is to simply use the direct java methods with proper
type hints when necessary.

Also I found out, that glfwGetCursorPos is incredibly slow. So the
polling solution went to the bin in favor of a callback.