small performance improvements
for some reason there was another map-indexed call in the exec function
of drawers. luckily it wasn't wrapped in a doall, so since that is a
lazy op, nothing really happened here, and the performance impact wasnt
that big, but it still was there.

also made a "lots of drawers" stress test for the "fastobj" demo.
With more and more drawers performance will be cpu-bound and we can
observe where exactly clsl spends most of its time.

initial testing revealed just a comparatively small amount of time
spent in glUseProgram. More time seems to be spent in uniform variable
setup. Especially glTexImage seems to dominate.